+++ app/server/src/server/framework-preset-server.ts
@@ -1,6 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Configuration } from 'webpack';
 import path from 'path';
+import { findDistEsm, StorybookConfig } from '@storybook/core-common';
 
 export function webpack(config: Configuration) {
   config.module.rules.push({
@@ -17,3 +18,7 @@ export function webpack(config: Configuration) {
 
   return config;
 }
+
+export const config: StorybookConfig['config'] = (entry = []) => {
+  return [...entry, findDistEsm(__dirname, 'client/preview/config')];
+};