+++ app/web-components/src/server/framework-preset-web-components.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Configuration } from 'webpack';
-import type { Options } from '@storybook/core-common';
+import { findDistEsm, Options, StorybookConfig } from '@storybook/core-common';
 
 export function webpack(config: Configuration, options: Options) {
   const babelrcOptions = options.features?.babelModeV7 ? null : { babelrc: false };
@@ -39,3 +39,7 @@ export function webpack(config: Configuration, options: Options) {
 
   return config;
 }
+
+export const config: StorybookConfig['config'] = (entry = []) => {
+  return [...entry, findDistEsm(__dirname, 'client/preview/config')];
+};