+++ examples/react-ts/.storybook/main.ts
@@ -1,13 +1,18 @@
 import type { StorybookConfig } from '@storybook/react/types';
 
 const config: StorybookConfig = {
-  stories: [{ directory: '../src', titlePrefix: 'Demo' }],
+  stories: [
+    {
+      directory: '../src',
+      titlePrefix: 'Demo',
+    },
+  ],
   logLevel: 'debug',
-  framework: '@storybook/react',
   addons: [
     '@storybook/addon-essentials',
     '@storybook/addon-storysource',
     '@storybook/addon-storyshots',
+    '@storybook/addon-a11y',
   ],
   typescript: {
     check: true,
@@ -18,6 +23,7 @@ const config: StorybookConfig = {
   },
   core: {
     builder: 'webpack4',
+    channelOptions: { allowFunction: false, maxDepth: 10 },
   },
   features: {
     postcss: false,
@@ -26,6 +32,6 @@ const config: StorybookConfig = {
     buildStoriesJson: true,
     babelModeV7: true,
   },
+  framework: '@storybook/react',
 };
-
 module.exports = config;