+++ lib/core-common/src/presets.ts
@@ -302,9 +302,13 @@ const getFrameworkPackage = (configDir: string) => {
   const main = serverRequire(resolve(configDir, 'main'));
   if (!main) return null;
   const { framework: frameworkPackage, features = {} } = main;
-  if ((features.storyStoreV7 || features.breakingChangesV7) && !frameworkPackage) {
+  if (features.breakingChangesV7 && !frameworkPackage) {
     throw new Error(dedent`
       Expected 'framework' in your main.js, didn't find one.
+
+      You can fix this automatically by running:
+
+      npx sb@next automigrate
     
       More info: https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#mainjs-framework-field
     `);